Implement plugin system (#11, PR #35)

+++ nyanyabot/database/migrations/0001.init.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS `bot_chats` (
 ) DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci COMMENT='NyaNyaBot chat data';
 
 CREATE TABLE IF NOT EXISTS `bot_users` (
-  `id` INT NOT NULL,
+  `id` bigint(20) NOT NULL,
   `first_name` VARCHAR(255) DEFAULT NULL,
   `last_name` VARCHAR(255) DEFAULT NULL,
   `username` VARCHAR(50) DEFAULT NULL,
@@ -18,7 +18,7 @@ CREATE TABLE IF NOT EXISTS `bot_users` (
 
 CREATE TABLE IF NOT EXISTS `lnk_bot_chats_id__bot_users_id` (
   `chat_id` bigint(20) NOT NULL,
-  `user_id` INT NOT NULL,
+  `user_id` bigint(20) NOT NULL,
   PRIMARY KEY (`chat_id`,`user_id`),
   CONSTRAINT `FK_chats_users_chats` FOREIGN KEY (`chat_id`) REFERENCES `bot_chats` (`id`),
   CONSTRAINT `FK_chats_users_users` FOREIGN KEY (`user_id`) REFERENCES `bot_users` (`id`)
